+++ _new_images_folder/game/illustrations/chapter1/illustration2/illustration2.php
@@ -1,0 +1,15 @@
+<?php
+
+include "../../../../../profile/imageGenerator.php";
+include "../../illustrationGenerator.php";
+
+$thisChapter = 1;
+$thisIllustration = 2;
+
+$destination = generateIllustration($thisChapter, $thisIllustration);
+
+// Display the final product
+header('Content-Type: image/png');
+imagepng($destination);
+
+?>